Add function 'update-pois'
The command 'update-pois' will create a json file with the updated data of all POIs. The file ivalidPOIs.txt includes a list of POIs that are present in decentralands catalyst API but must not be show on the map. The file missingNames.js includes the names of POIs that are not incuded in the API response.